Organized components better with non-routable folders and started dynamic api routes for nextjs
