+++ src/app/marketing/_components/FAQ.tsx
