+++ src/app/marketing/_components/Hero.tsx
