+++ src/app/marketing/page.tsx
@@ -1,11 +1,11 @@
 
 import { Divider } from "@mui/material";
-import Hero from "@/components/Hero";
-import LogoCollection from "@/components/LogoCollection";
-import Features from "@/components/Features";
-import Highlights from "@/components/Highlights";
-import Pricing from "@/components/Pricing";
-import FAQ from "@/components/FAQ";
+import Hero from "@/app/marketing/_components/Hero";
+import LogoCollection from "@/app/marketing/_components/LogoCollection";
+import Features from "@/app/marketing/_components/Features";
+import Highlights from "@/app/marketing/_components/Highlights";
+import Pricing from "@/app/marketing/_components/Pricing";
+import FAQ from "@/app/marketing/_components/FAQ";
 
 export default function Home() {
   return (
